feat(persistence-errors): Use GetAccountId error
This baloons into a big change, since the method is used in multiple
places. But it shows how the error hierarchy actually is.

As a method signature is changed, including just changing the error type
the depending methods need to be given consideration. In terms of how they
will handle error variants from the error returning method.

Signed-off-by: Naian <126972030+nain-F49FF806@users.noreply.github.com>